updated the codes of ExportButton, FileImport, useExcelData and excelUtils. also downloaded missing dependencies for the chart.
