+++ src/components/FileImport.tsx
@@ -1,57 +1,13 @@
 import { Button, Box } from "@mui/material";
-import { parseExcelFile } from "../utils/excelUtils";
-import type { TableData, SpaceMineData } from "../types/Types";
-import { useTableData } from "../hooks/useTableData";
+import { useExcelData } from "../hooks/useExcelData";
+import type { TableData } from "../types/Types";
 
 type FileImportProps = {
   onDataLoaded: (data: TableData) => void;
 };
 
 export default function FileImport({ onDataLoaded }: FileImportProps) {
-  const { tableData, setTableData } = useTableData();
-
-  const handleFileUpload = async (
-    event: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    const file = event.target.files?.[0];
-    if (!file) return;
-
-    try {
-      const parsedData = await parseExcelFile(file);
-      setData(parsedData);
-      onDataLoaded(parsedData);
-    } catch (error) {
-      console.error("Fel Vid uppladdning av Excel:", error);
-    }
-    const workbook = new ExcelJS.Workbook();
-    const arrayBuffer = await file.arrayBuffer();
-    await workbook.xlsx.load(arrayBuffer);
-
-    const worksheet = workbook.getWorksheet(1);
-    if (!worksheet) {
-      console.error("Inget första ark hittades i filen!");
-      return;
-    }
-
-    const newData: SpaceMineData[] = [];
-
-    worksheet.eachRow((row, rowNumber) => {
-      if (rowNumber === 1) return; // Skippa header-raden
-
-      const astronaut = row.getCell(1).value?.toString().trim() || "";
-      const planet = row.getCell(2).value?.toString().trim() || "";
-      const mineral = row.getCell(3).value?.toString().trim() || "";
-      const amount = Number(row.getCell(4).value) || 0;
-
-      // Bara lägg till om det finns data
-      if (astronaut && planet && mineral) {
-        newData.push({ astronaut, planet, mineral, amount });
-      }
-    });
-
-    setTableData(newData);
-    onDataLoaded(newData);
-  };
+  const { tableData, handleFileUpload } = useExcelData(onDataLoaded);
 
   return (
     <Box
